+++ base64.h
@@ -22,6 +22,7 @@
 #define BASE64_H
 
 #include <stdint.h>
+#include <stddef.h>
 
 /**
  * Decode a base64-encoded string.
@@ -33,7 +34,7 @@
  * @return         number of bytes written, or a negative value in case of
  *                 invalid input
  */
-int base64_decode(uint8_t *out, const char *in, int out_size);
+size_t base64_decode(uint8_t *out, const char *in, int out_size);
 
 /**
  * Encode data to base64 and null-terminate.
@@ -44,7 +45,7 @@ int base64_decode(uint8_t *out, const char *in, int out_size);
  * @param in_size  size in bytes of the 'in' buffer
  * @return         'out' or NULL in case of error
  */
-char *base64_encode(char *out, int out_size, const uint8_t *in, int in_size);
+char *base64_encode(char *out, size_t out_size, const uint8_t *in, size_t in_size);
 
 /**
  * Calculate the output size needed to base64-encode x bytes.
